Django/Backend/requirements.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
drf-spectacular>=0.15.1,<0.16
Unidecode>=1.3.4,<1.4
more-itertools>=8.13.0,<8.14
openpyxl>=3.0.10,<3.1
pandas
pdfservices-sdk
xlrd>=2.0.1,<2.1
XlsxWriter>=3.0.3,<3.1
django-cleanup>=6.0.0,<6.1
django-filter>=22.1,<22.2
mssql-django>=1.1.3,<1.1.4
django-cors-headers>=3.13.0
django-environ>=0.9.0
gunicorn>=20.1.0
environ